backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
websockets==12.0
solders==0.20.0
base58==2.1.1
python-dotenv==1.0.1
pydantic==2.6.0
